DERBY-2896
Backporting changes 557334 from main into 10.3 codeline. The commit comments for 557334 were as follows

Metadata calls getTables and getUDTs were failing when run from a user schema in a territory based collated database. 
The reason for it is that these metadata calls were not getting compiled in SYS schema when they were executed from 
a user schema. Metadata calls should always compile in SYS schema no matter what the current schema might be. The 
reason getTables was not getting compiled in SYS schema was because we were trying to modify it's metadata sql on 
the fly and then were compiling that modified sql in whatever the current schema might be. I have changed the 
metadata sql for getTables in metadata.properties so that we do not need to modify it on the fly anymore. This will 
allow getTables to follow the same codepath as other metadata queries which will also ensure that the sql gets 
compiled in SYS schema. 

As for getUDTs, it was merely a coding bug that we didn't follow the same logic as other metadata queries for it. 
I have changed getUDTs implementation to follow the same codepath as other metadata queries. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@557405 13f79535-47bb-0310-9956-ffa450edef68