Migrating changes 557886 for DERBY-2955 from main into 10.3 codeline. The commit comments in main were as follows
We used to set the collation type of character string columns in the generate phase rather than the bind phase of create table. But this will cause problem with following query
       CREATE TABLE STAFF9 (EMPNAME CHAR(20),
       CONSTRAINT STAFF9_EMPNAME CHECK (EMPNAME NOT LIKE 'T%'))
For the query above, when run in a territory based db, we need to have the correct collation set in bind phase of create table so that when LIKE is handled in LikeEscapeOperatorNode, we have the correct collation set for EMPNAME otherwise it will throw an exception for 'T%' having collation of territory based and EMPNAME having the default collation of UCS_BASIC. The change in this commit will ensure that character string columns get their collation set early on in the bind phase so when the bind code for LIKE kicks in, we are all set with correct collation information.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@557910 13f79535-47bb-0310-9956-ffa450edef68