DERBY-2887: NULLS FIRST/LAST for ORDER BY
This change implements the SQL Standard 10.10 Sort Specification List feature:

  <null ordering :== NULLS FIRST | NULLS LAST

The implementation adds a new boolean argument to the DataType.compare()
function to allow control over whether NULL values should be compared
as lower than non-NULL values or as higher. The change also adds code to
the parser to recognize the new syntax, and to pass it to the execution
layer via the compiler data structures.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@567314 13f79535-47bb-0310-9956-ffa450edef68