Migrating comments changes from main (revision 570482) into 10.3 codeline. The commit comments in main were as follows
Fixed the comments for getDominantTypeServices() method in ValueNodeList. The comments were using an example of concatenation operation but concatenation operation does not call this method to find the dominant type of it's 2 operands. Instead, concatenation operator implements it's own simple algorithm in ConcatenationOperatorNode. This is because ConcatenationOperatorNode does not work off of a ValueNodeList. I have reworded the comments for getDominantTypeServices() to describe what is doing.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@570483 13f79535-47bb-0310-9956-ffa450edef68