DERBY-3085 Fails to handle BLOB fields with a PreparedStatement with size >32750 bytes
Store a reference to the stream for the streamed parameter in the DRDAStatement.paramState and then drain the stream after statement execution if needed. There is only one parameter ever streamed, so only one field needed to be added. I added a test for both BLOB's and CLOB's to Blob4ClobTest.java.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@581012 13f79535-47bb-0310-9956-ffa450edef68