Migrating changes from trunk(586596) for DERBY-3046 into 10.3 codeline with this commit. The commit changes for trunk were as follows
This commit fixes the null pointer exception by making sure that we do not try to access the object's elements if the object is null. I have also added a test case for this fix.

Additionally, while working on this, I found that we return different sql state for parameter being out of range for PreparedStatment.setObject and Preparedstatement.setXXX calls in Network Server. I have entered a Jira entry for that issue DERBY-3139.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@586607 13f79535-47bb-0310-9956-ffa450edef68