DERBY-3054
The non-junit test, synonym.sql, tests the synonym dependency. The particular test case 2 views dependent on it. 
Depending on the jvm being used, the 2 exceptions(with same sql state but different object dependency in the sql
message text) thrown for dependency might be in different order and that causes test failure. In order to fix this,
I have carved out that test from synonym.sql into a new junit test and since in junit test, we can simply assert
the sql state, the order of 2 exceptions does not matter. I will migrate this fix into Derby 10.3 soon.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@587553 13f79535-47bb-0310-9956-ffa450edef68