Merging change (revision 588147) from trunk into 10.3 codeline. The commit comments for trunk were as follows DERBY-3136
Cut down on unnecessary String and CollationElementIterator objects creation in Like.java class by changing the way
we compare the character in value string vs character in the pattern string.

When working with database with UCS_BASIC collation, the 2 characters should be the same otherwise we will return
FALSE.

When working with territory based collation, if the 2 characters are the same, then we can simply return TRUE
because their collation elements will have to be same. But if the 2 characters are not same, then use the
RuleBasedCollator for the territory to see if those 2 different characters match (ie they have same collation
element(s) corresponding to them).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@588148 13f79535-47bb-0310-9956-ffa450edef68