DERBY-766 (partial) Add code to the byte code compiler to enable it to analyze a block of existing generated byte code. Provide utility methods to allow other methods to be written to walk the byte code. Add an method to find the maximum stack depth of a piece of java code.
Add sanity checks to double check the data used (OPCODE_ACTION) to walk the byte code and compare the calculated max stack depth with the existing generated value. This is progress towards providing the ability to split generated methods that exceed the JVM's limit into a number of smaller methods. This will replace the existing mechanisms which are reactive, non-optimal and only work in a limited number of situations.

merge from trunk revision 376882



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@590719 13f79535-47bb-0310-9956-ffa450edef68