DERBY-3175: NullPointerException on INSERT after ALTER TABLE DROP COLUMN
The implementation of ALTER TABLE DROP COLUMN needs to adjust the
SYSCOLUMNS information for some of the columns in the table. Specifically,
columns with a higher column position than the dropped column need to
have their column position adjusted (decremented by 1).

The algorithm which did this column position adjustment was flawed,
because it was losing the auto-increment information for generated columns.

The code adjusts the column position by:
 - reading the old column information into a ColumnDescriptor
 - deleting the old column information from SYSCOLUMNS
 - modifying the ColumnDescriptor to set the new column position
 - adding the new column information to SYSCOLUMNS

But the generated column information in the ColumnDescriptor was not being
preserved properly in this process.

The fix involved ensuring that the ColumnDescriptor information was
accurate for generated columns, and ensuring that SYSCOLUMNSRowFactory.makeRow
was correctly loading the generated column information into the new row.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@594727 13f79535-47bb-0310-9956-ffa450edef68