DERBY-3172
Change the Network Server code so that when the Network Server is shutdown, any api on Connection object will 
raise a Connection error event (similar to what is done in embedded mode). This also required me changing the
disconnectError from -4999 to 40000(to match embedded). This change was made client.am.SqlCode. I have added tests
for all the apis on Connection object in DataSourceTest. Had to change couple tests which were testing for
-4999. Now they look for 40000. I will merge this change into 10.3 and run tests there before committing the changes
in 10.3 codeline.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@595047 13f79535-47bb-0310-9956-ffa450edef68