DERBY-2352: Assertion failure with order by and group by on SUBSTR expression
For a query plan which involves a sort, the compilation system generates a
method called a "row allocator" which allocates a template row of the
correct shape (correct number of columns of the correct data types) to hold
the row that is being provided to the sorter. The row allocator determines
the data type of the column in the row allocator by examining the 
ResultColumn's type, which in turn is computed by the compilation system
during bind processing.

For the SUBSTR expression, TernaryOperatorNode.substrBind() is called, and
the logic in that method was determining that the data type of the SUBSTR
result depended on the data type of the operand on which SUBSTR was called;
i.e., a SUBSTR of a CHAR became a CHAR, while a SUBSTR of a VARCHAR became
a VARCHAR. But this is incorrect; SUBSTR always returns a VARCHAR type,
unless it is called on a BLOB/CLOB object.

This patch modifies the substrBind() processing to perform result data type
analysis in the same way as is done for the trim family of methods in trimBind.





git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@596824 13f79535-47bb-0310-9956-ffa450edef68