Merging change(599933) from 10.3 codeline into trunk for DERBY-3222 The commit comment for checkin into 10.3 codeline was as follows
The test will fail depending on the timing. It inserts CURRENT_DATE into a column and then compares that
column with whatever the CURRENT_DATE might be during the run of compare sql. If the insert happened right before
midnight and comparison happens right after midnight, the comparison of the dates will return FALSE. Since the main 
purpose of the test is to make sure that Derby derives correct collation type for CURRENT_DATE and 
UPPER(CURRENT_DATE), it really does not matter if the 2 current dates actually match or not. Because of this, I 
have changed the test to look for <= while doing date comparison rather than =. This should resolve the timing 
related fact of the test.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@599973 13f79535-47bb-0310-9956-ffa450edef68