DERBY-3238 When table contains large LOB values (> ~32K) trigger execution fails for that row with ERROR XCL30: An IOException was thrown when reading a 'BLOB'
When a trigger is present, unchanged columns with be part of the UpdateResultSet. These columns are present twice in the result set as a before and after value. There was code to "objectify" the before value but the after value was not getting updated. Changed objectifyStream to make sure after values point to data value and not the stream after objectifyStream occurs.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@601110 13f79535-47bb-0310-9956-ffa450edef68