Migrating change(601830) from trunk into 10.3 codeline for DERBY-3037. The commit comments for trunk were as follows
DERBY-3037

This commit makes sure that EmbeddedResultSet.close() calls Language Resultset.close rather than Language
Resultset.finish. This change caused few tests to fail. The tests had runtime statistics on but the code to dump
the runtime statistics was in Language Resutlset.finish and not in Language Resultset.close. To fix this, I have
moved the code to dump runtime statistics from Language Resutlset.finish into Lanugage ResultSet.close This has
fixed the test failures.

As for the 2nd part of this jira entry which is to have activation.close() call Language Resultset.finish(). This
already happens in impl.sql.execute.BaseActivation.close() and hence no work was needed for the 2nd part.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@602198 13f79535-47bb-0310-9956-ffa450edef68