DERBY-3261 and part of DERBY-3037
The outerjoin.sql was failing because the part of the runtime statistcis info was getting erased before
LanguageResultSet.close() code collected it all. I moved the erasing of runtime stat code so that it happened
once the stat was collected successfully.

In addition, I removed redundant code of closing and finishing the LanguageResultSet from EmbedResultSet.java
because these steps happen as part of activation.close

I will merge this into 10.3 codeline and fire the tests there.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@603823 13f79535-47bb-0310-9956-ffa450edef68