Merging changes from trunk(603823) into 10.3 codeline for DERBY-3261 and part of DERBY-3037
The commit comments for trunk were as follows

The outerjoin.sql was failing because the part of the runtime statistcis info was getting erased before LanguageResultSet.close() code collected it all. I moved the erasing of runtime stat code so that it happened once the stat was collected successfully. 

In addition, I removed redundant code of closing and finishing the LanguageResultSet from EmbedResultSet.java because these steps happen as part of activation.close 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@603980 13f79535-47bb-0310-9956-ffa450edef68