Adding a test case for DERBY-3302. There are two sub-test case classes involved here, col_rec1.java and col_rec2.java
The purpose of col_rec1 test and col_rec2 test is to create a territory based database and 
create some objects with collation sensitive character types. Then, make the database crash 
so that during the recovery, store engine has to do collation related operations. Those 
collation related operations are going to require that we use correct Collator object. 
DERBY-3302 demonstrated a npe during this operation because Derby was relying on database 
context to get the correct Collator object. But database context is not available at this 
point in the recovery. With the fix for DERBY-3302, the Collator object will now be 
obtained from collation sensitive datatypes itself rather than looking at database context 
which is not available at this point in recovery. 

col_rec1 class will do the steps of create a territory based database and create some 
objects with collation sensitive character types. Then, make the database crash. 
col_rec2 class will do the part of rebooting the crashed db which will require store to go 
through recovery.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@611295 13f79535-47bb-0310-9956-ffa450edef68