Merging changes 610846 from trunk into 10.3 codeline. This is for DERBY-3302. Commit comments for trunk 610846 were as follows
DERBY-3302 The user was running into null pointer exception at the time of database recovery 
because Derby was trying to get the Collator object through database context. But the 
Collator object is already available in the territory sensitive character classes and we 
do not have to go to database context to get it. I changed the code to use that collator 
object rather than look into database context. The reason for null pointer exception was 
that database context was not loaded yet during database recovery. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@611297 13f79535-47bb-0310-9956-ffa450edef68