Merging changes from trunk(revision 612865) into 10.3 codeline to fix the failure in SetObjectUnsupportedTest. The commit comments for trunk were as follows
While working on DERBY-2720, I removed national datatypes as one of the unsupported data
types(revision 612525) and that caused the test failure for SetObjectUnsupportedTest since
we stopped recognizing national dataypes as one of the unsupported datatypes. The fix
required changes in only one class which I had modified incorrectly in revision 612525.
This test failure was not caught earlier by my local runs of the test because I had used
JDK14 which has JDBC3 support and this paritcular test requires JDBC4 support.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@612872 13f79535-47bb-0310-9956-ffa450edef68