DERBY-3068 testImportExportProcedureNegative(org.apache.derbyTesting.functionTests.tests.tools.ImportExportProcedureTest)junit.framework.ComparisonFailure: Unexpected SQL state. expected:<38000> but was:<XIE0S>
Changed unexpectedException in import to close the stream first.
Changed ImportExportProcedureTest to delete files immediately after the test instead of before the test that recreates the file.
Changed deleteFile to fail if it cannot delete the file.

Ran suites.All on IBM 1.5 3 times without errors, and on jdk16.  There may however still be a problem with import as I did see the unchanged test fail on IBM 1.5. With the new test structure it should be easier to identify the source of the problem if it happens again.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@614893 13f79535-47bb-0310-9956-ffa450edef68