Merging revision 627673 from trunk into 10.3 codeline. The commit comments for 627673 were as follows
DERBY-3304 and DERBY-3414

This serves as a test case for both the jira entries above.

Number of code changes for transaction ending time went in as part of DERBY-3304 and this
new test will check those code changes for specific case of rollback inside a java procedure
call.

The test case is currently disabled for network server because the rollback inside the
procedure is not closing all the resultsets and DERBY-3414 is to track this behavior of
network server. Once DERBY-3414 is fixed, we should enable the test for network server.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@627674 13f79535-47bb-0310-9956-ffa450edef68