DERBY-3163: Derby JUnit test framework adaptions to run existing tests against replicated databases Contributed by Ole Solberg.
- TestConfiguration:
Add decorators which takes server host and port, for running tests in a
server/client configuration, where a server has already
been started on a that host and port.

- CleanDatabaseTestSetup:
Added a constructor which takes server host and port, for an already started server in a
client/server configuration, to run the SQL decoration on that server.

- AnsiTrimTest, CreateTableFromQueryTest:
Factored out the SQL statement execution for database content setup to
be able to reuse the same code in other TestCases which add the same
test method in their 'suite()' method. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@628073 13f79535-47bb-0310-9956-ffa450edef68