DERBY-3304 DERBY-3037 DERBY-1585
I am adding a test case to check for the resultset from the java procedure call when the
java procedure has done a rollback inside it. This test shows that in trunk, after checkin 
602991 for DERBY-1585, a procedure does not return a resultset if there was a rollbck 
inside the procedure with resultset creation before rollback. This behavior is different 
than what happens in 10.2 codeline. In 10.2, a procedure will return a *closed* resultset 
if there was a rollback inside the procedure. But a procedure should not return closed 
result sets, so it appears that trunk is behaving correctly and 10.2's behavior was 
incorrect.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@628130 13f79535-47bb-0310-9956-ffa450edef68